+++ Krowi_AchievementFilter/GUI/WorldMapButton.lua
@@ -9,7 +9,9 @@ local worldMapButton = gui.WorldMapButton;
 addon.WorldMapButtons = LibStub("Krowi_WorldMapButtons-1.0"); -- Global world map buttons object
 
 function worldMapButton.Load()
-    addon.WorldMapButtons:Add("KrowiAF_WorldMapAchievementButtonTemplate", "BUTTON");
+    worldMapButton = addon.WorldMapButtons:Add("KrowiAF_WorldMapAchievementButtonTemplate", "BUTTON");
+
+    return worldMapButton;
 end
 
 WorldMapAchievementButtonMixin = {};
@@ -31,17 +33,39 @@ function WorldMapAchievementButtonMixin:OnMouseUp()
 end
 
 function WorldMapAchievementButtonMixin:OnClick()
-    
+    local categoriesFrame = gui.GetFrames("CategoriesFrame");
+    if worldMapButton.Achievements and #worldMapButton.Achievements > 0 then
+        addon.GUI.ToggleAchievementFrameAtTab1(true);
+        addon.Categories[2].Name = addon.L["Selected Zone"] .. " (" .. worldMapButton.name .. ")";
+        categoriesFrame:SelectCategory(addon.Categories[2]);
+    end
 end
 
 function WorldMapAchievementButtonMixin:OnEnter()
-    GameTooltip:SetOwner(self, "ANCHOR_RIGHT");
-    GameTooltip_SetTitle(GameTooltip, AF_NAME);
+    if not IsAddOnLoaded("Blizzard_AchievementUI") then
+        LoadAddOn("Blizzard_AchievementUI");
+    end
 
-    local refresh = addon.Data.GetCurrentZoneAchievements();
+    local mapID = WorldMapFrame:GetMapID();
+    addon.SelectedZoneCategory.Achievements = addon.GetAchievementsInZone(mapID);
+    -- replace this by a new category named SelectedZone
+    local numOfAch, numOfCompAch, numOfNotObtAch = 0, 0, 0;
+    for _, achievement in next, addon.SelectedZoneCategory.Achievements do
+        numOfAch, numOfCompAch, numOfNotObtAch = addon.Validate(worldMapButton, achievement, numOfAch, numOfCompAch, numOfNotObtAch); -- , numOfIncompAch
+    end
 
-    -- This zone has x completed, y not completed and z not obtainable achievements.
-    -- Click me to show a list of not completed, completed and not obtainable achievements in this order.
+    worldMapButton.Achievements = addon.SelectedZoneCategory.Achievements;
+    worldMapButton.name = C_Map.GetMapInfo(mapID).name;
+    worldMapButton.numAchievements = numOfAch;
+    worldMapButton.numCompleted = numOfCompAch;
+    worldMapButton.numOfNotObtAch = numOfNotObtAch;
+    local numOfNotObtAchText = "";
+    if numOfNotObtAch > 0 and addon.Options.db.Tooltip.Categories.ShowNotObtainable then
+        numOfNotObtAchText = " (+" .. numOfNotObtAch .. ")";
+    end
+    worldMapButton.numCompletedText = numOfCompAch .. numOfNotObtAchText .. " / " .. numOfAch;
+
+    addon.StatusBarTooltip(worldMapButton);
 end
 
 function WorldMapAchievementButtonMixin:OnHide()